# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shop API
  version: 1.0.0
  description: |
    REST API магазина бытовой техники для управления клиентами, товарами,
    поставщиками и изображениями. Все идентификаторы имеют формат UUID.

    Все методы `/api/v1`, кроме регистрации, входа (в том числе через
    OAuth-провайдера) и сброса пароля, требуют access-токен в заголовке
    `Authorization: Bearer <token>`. Токен выдают методы
    `POST /api/v1/auth/register`, `POST /api/v1/auth/login` и
    `POST /api/v1/auth/oauth/callback`, а проверяет его сервис авторизации.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: Локальный сервер разработки
    variables:
      port:
        default: '8080'
        description: Порт HTTP-сервера приложения
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Сменить пароль
      description: |
        Меняет пароль пользователя, которому принадлежит access-токен, и
        выдаёт новый токен.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              old_password: s3cret-pass
              new_password: n3w-s3cret-pass
      responses:
        '200':
          description: Пароль изменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            Access-токен отсутствует или недействителен (`UNAUTHORIZED`) либо
            старый пароль неверен (`INVALID_CREDENTIALS`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_CREDENTIALS
                message: old password is incorrect
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Сбросить пароль
      description: |
        Назначает пользователю временный пароль. Ответ одинаков независимо
        от того, существует ли учётная запись с таким email, чтобы по нему
        нельзя было выяснить список пользователей.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              email: ivan@example.com
      responses:
        '204':
          description: Запрос на сброс пароля принят
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/oauth/{provider}/start:
    post:
      tags: [Auth]
      summary: Начать вход через OAuth-провайдера
      description: |
        Возвращает адрес страницы авторизации провайдера. Фронтенд отправляет
        туда пользователя; после входа провайдер вернёт его на настроенный
        redirect URI с параметрами `code` и `state`, которые нужно передать в
        `POST /api/v1/auth/oauth/callback`. Адрес действителен ограниченное
        время (по умолчанию 10 минут).
      operationId: oauthStart
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          description: Идентификатор провайдера
          schema:
            type: string
            enum: [yandex]
      responses:
        '200':
          description: Адрес страницы авторизации провайдера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthStartResponse'
        '400':
          description: Провайдер не поддерживается или не настроен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: VALIDATION_ERROR
                message: unsupported oauth provider
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/oauth/callback:
    post:
      tags: [Auth]
      summary: Завершить вход через OAuth-провайдера
      description: |
        Обменивает код авторизации на профиль пользователя у провайдера и
        выдаёт access-токен. Провайдер определяется по `state`. Если
        пользователя ещё нет, он создаётся без пароля; если есть пользователь
        с тем же email, подтверждённым провайдером, учётная запись провайдера
        привязывается к нему.
      operationId: oauthCallback
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCallbackRequest'
            example:
              code: '4723981'
              state: Zq3H2l0x9vYc1P6nT8wKfR5aBdE7uJmS4gN0iLoXyWc
      responses:
        '200':
          description: Вход выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            `state` неизвестен или просрочен, код авторизации недействителен
            либо провайдер не вернул email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: OAUTH_FAILED
                message: oauth authorization failed
        '409':
          description: |
            Пользователь с таким email уже существует, а провайдер не
            подтвердил владение адресом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: USER_ALREADY_EXISTS
                message: user with this email already exists
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из ответа регистрации или входа
...
  responses:
...
    BadRequest:
      description: Некорректный запрос или ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: validation error
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: resource not found
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: internal server error
  schemas:
...
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          format: password
          minLength: 1
          example: s3cret-pass
        new_password:
          type: string
          format: password
          minLength: 1
          example: n3w-s3cret-pass
    ResetPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: ivan@example.com
    OAuthStartResponse:
      type: object
      required: [auth_url]
      properties:
        auth_url:
          type: string
          format: uri
          description: Адрес страницы авторизации провайдера
          example: https://oauth.yandex.ru/authorize?response_type=code&client_id=...&state=...
    OAuthCallbackRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
          minLength: 1
          description: Код авторизации из redirect URI
        state:
          type: string
          minLength: 1
          description: Значение state из redirect URI
    OAuthLoginResponse:
      type: object
      required: [user_id, email, token]
      properties:
        user_id:
          type: string
          format: uuid
          example: 3b1c2f0e-6a7d-4f4e-9d61-2a3c8b9e5f10
        email:
          type: string
          format: email
          example: ivan@yandex.ru
        token:
          $ref: '#/components/schemas/Token'
    Token:
      type: object
      required: [access_token, expires_at]
      properties:
        access_token:
          type: string
          description: 'JWT для заголовка `Authorization: Bearer <token>`'
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_at:
          type: string
          format: date-time
          example: '2026-09-14T10:30:00Z'
...
    ChangePasswordResponse:
      type: object
      required: [token]
      properties:
        token:
          $ref: '#/components/schemas/Token'
    Error:
      type: object
      description: Код ошибки для обработки на фронтенде и человекочитаемое сообщение
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: VALIDATION_ERROR
        message:
          type: string
          description: Человекочитаемое сообщение об ошибке
          example: validation error